organize code roughly, create jaxmodulewrapper
