+++ scvi/model/_jaxscvi.py
@@ -1,36 +1,25 @@
 import logging
-from typing import Any, Optional, Sequence, Union
+from typing import Optional, Sequence
 
 import jax
 import jax.numpy as jnp
 import numpy as np
-import optax
-import pandas as pd
-import tqdm
 from anndata import AnnData
-from flax.core import FrozenDict
-from flax.training import train_state
-from jax import random
 
 from scvi import REGISTRY_KEYS
 from scvi._compat import Literal
 from scvi.data import AnnDataManager
 from scvi.data.fields import CategoricalObsField, LayerField
-from scvi.dataloaders import DataSplitter
 from scvi.module import JaxVAE
-from scvi.train import JaxModuleInit, JaxTrainingPlan, TrainRunner
+from scvi.module.base import JaxModuleWrapper
 from scvi.utils import setup_anndata_dsp
 
-from .base import BaseModelClass
+from .base import BaseModelClass, JaxTrainingMixin
 
 logger = logging.getLogger(__name__)
 
 
-class TrainState(train_state.TrainState):
-    batch_stats: FrozenDict[str, Any]
-
-
-class JaxSCVI(BaseModelClass):
+class JaxSCVI(BaseModelClass, JaxTrainingMixin):
     """
     EXPERIMENTAL single-cell Variational Inference [Lopez18]_, but with a Jax backend.
 
@@ -76,7 +65,8 @@ class JaxSCVI(BaseModelClass):
 
         n_batch = self.summary_stats.n_batch
 
-        self.module_kwargs = dict(
+        self.module = JaxModuleWrapper(
+            JaxVAE,
             n_input=self.summary_stats.n_vars,
             n_batch=n_batch,
             n_hidden=n_hidden,
@@ -84,9 +74,8 @@ class JaxSCVI(BaseModelClass):
             dropout_rate=dropout_rate,
             is_training=False,
             gene_likelihood=gene_likelihood,
+            **model_kwargs,
         )
-        self.module_kwargs.update(model_kwargs)
-        self._module = None
 
         self._model_summary_string = ""
         self.init_params_ = self._get_init_params(locals())
@@ -119,271 +108,199 @@ class JaxSCVI(BaseModelClass):
         adata_manager.register_fields(adata, **kwargs)
         cls.register_manager(adata_manager)
 
-    def _get_module(self, kwargs=None):
-        if kwargs is None:
-            kwargs = self.module_kwargs
-        return JaxVAE(**kwargs)
+    # def train_manual(
+    #     self,
+    #     max_epochs: Optional[int] = None,
+    #     check_val_every_n_epoch: Optional[int] = None,
+    #     use_gpu: Optional[Union[str, int, bool]] = None,
+    #     train_size: float = 0.9,
+    #     validation_size: Optional[float] = None,
+    #     batch_size: int = 128,
+    #     lr: float = 1e-3,
+    # ):
+    #     """
+    #     Train the model.
 
-    @property
-    def module(self):
-        if self._module is None:
-            self._module = self._get_module()
-        return self._module
+    #     Parameters
+    #     ----------
+    #     max_epochs
+    #         Number of passes through the dataset. If `None`, defaults to
+    #         `np.min([round((20000 / n_cells) * 400), 400])`
+    #     use_gpu
+    #         Use default GPU if available (if None or True), or index of GPU to use (if int),
+    #         or name of GPU (if str, e.g., `'cuda:0'`), or use CPU (if False).
+    #     train_size
+    #         Size of training set in the range [0.0, 1.0].
+    #     validation_size
+    #         Size of the test set. If `None`, defaults to 1 - `train_size`. If
+    #         `train_size + validation_size < 1`, the remaining cells belong to a test set.
+    #     batch_size
+    #         Minibatch size to use during training.
+    #     """
+    #     if max_epochs is None:
+    #         n_cells = self.adata.n_obs
+    #         max_epochs = np.min([round((20000 / n_cells) * 400), 400])
 
-    def train(
-        self,
-        max_epochs: Optional[int] = None,
-        use_gpu: Optional[Union[str, int, bool]] = None,
-        train_size: float = 0.9,
-        validation_size: Optional[float] = None,
-        batch_size: int = 128,
-        lr: float = 1e-3,
-        **trainer_kwargs,
-    ):
-        if max_epochs is None:
-            n_cells = self.adata.n_obs
-            max_epochs = np.min([round((20000 / n_cells) * 400), 400])
+    #     data_splitter = DataSplitter(
+    #         self.adata_manager,
+    #         train_size=train_size,
+    #         validation_size=validation_size,
+    #         batch_size=batch_size,
+    #         # for pinning memory only
+    #         use_gpu=False,
+    #         iter_ndarray=True,
+    #     )
+    #     data_splitter.setup()
+    #     train_loader = data_splitter.train_dataloader()
+    #     val_dataloader = data_splitter.val_dataloader()
 
-        data_splitter = DataSplitter(
-            self.adata_manager,
-            train_size=train_size,
-            validation_size=validation_size,
-            batch_size=batch_size,
-            # for pinning memory only
-            use_gpu=False,
-            iter_ndarray=True,
-        )
+    #     if val_dataloader is None:
+    #         raise UserWarning(
+    #             "No observations in the validation loop. No validation metrics will be recorded."
+    #         )
 
-        module_kwargs = self.module_kwargs.copy()
-        module_kwargs.update(dict(is_training=True))
-        train_module = self._get_module(module_kwargs)
+    #     module_kwargs = self.module_kwargs.copy()
+    #     module_kwargs.update(dict(is_training=True))
+    #     train_module = self._get_module(module_kwargs)
 
-        # has is_training set to False
-        validation_module = self.module
+    #     # if key is generated on CPU, model params will be on CPU
+    #     # we have to pay the price of a JIT compilation though
+    #     if use_gpu is False:
+    #         key = jax.jit(lambda i: random.PRNGKey(i), backend="cpu")
+    #     else:
+    #         # dummy function
+    #         def key(i: int):
+    #             return random.PRNGKey(i)
 
-        self.training_plan = JaxTrainingPlan(
-            train_module, validation_module, use_gpu=use_gpu
-        )
-        self.training_plan.set_rngs(["params", "dropout", "z"])
-        if "callbacks" not in trainer_kwargs.keys():
-            trainer_kwargs["callbacks"] = []
-        trainer_kwargs["callbacks"].append(JaxModuleInit())
+    #     self.rngs = {
+    #         "params": key(0),
+    #         "dropout": key(1),
+    #         "z": key(2),
+    #     }
+    #     module_init = train_module.init(self.rngs, next(iter(train_loader)))
+    #     params = module_init["params"]
+    #     batch_stats = module_init["batch_stats"]
 
-        runner = TrainRunner(
-            self,
-            training_plan=self.training_plan,
-            data_splitter=data_splitter,
-            max_epochs=max_epochs,
-            use_gpu=False,
-            **trainer_kwargs,
-        )
-        runner()
+    #     optimizer = optax.chain(
+    #         optax.additive_weight_decay(weight_decay=1e-6),
+    #         optax.adam(lr, eps=0.01),
+    #     )
 
-        self.train_state = self.training_plan.state
-        self.params = self.train_state.params
-        self.batch_stats = self.train_state.batch_stats
-        self.is_trained_ = True
+    #     state = TrainState.create(
+    #         apply_fn=train_module.apply,
+    #         params=params,
+    #         tx=optimizer,
+    #         batch_stats=batch_stats,
+    #     )
 
-        self.module_kwargs.update(dict(is_training=False))
-        self._module = None
-        self.bound_module = self.module.bind(
-            {"params": self.params, "batch_stats": self.batch_stats},
-            rngs=self.training_plan.rngs,
-        )
+    #     @jax.jit
+    #     def train_step(state, array_dict, rngs, **kwargs):
+    #         rngs = {k: random.split(v)[1] for k, v in rngs.items()}
 
-    def train_manual(
-        self,
-        max_epochs: Optional[int] = None,
-        check_val_every_n_epoch: Optional[int] = None,
-        use_gpu: Optional[Union[str, int, bool]] = None,
-        train_size: float = 0.9,
-        validation_size: Optional[float] = None,
-        batch_size: int = 128,
-        lr: float = 1e-3,
-    ):
-        """
-        Train the model.
+    #         # batch stats can't be passed here
+    #         def loss_fn(params):
+    #             vars_in = {"params": params, "batch_stats": state.batch_stats}
+    #             outputs, new_model_state = state.apply_fn(
+    #                 vars_in, array_dict, rngs=rngs, mutable=["batch_stats"], **kwargs
+    #             )
+    #             loss_recorder = outputs[2]
+    #             loss = loss_recorder.loss
+    #             elbo = jnp.mean(
+    #                 loss_recorder.reconstruction_loss + loss_recorder.kl_local
+    #             )
+    #             return loss, (elbo, new_model_state)
 
-        Parameters
-        ----------
-        max_epochs
-            Number of passes through the dataset. If `None`, defaults to
-            `np.min([round((20000 / n_cells) * 400), 400])`
-        use_gpu
-            Use default GPU if available (if None or True), or index of GPU to use (if int),
-            or name of GPU (if str, e.g., `'cuda:0'`), or use CPU (if False).
-        train_size
-            Size of training set in the range [0.0, 1.0].
-        validation_size
-            Size of the test set. If `None`, defaults to 1 - `train_size`. If
-            `train_size + validation_size < 1`, the remaining cells belong to a test set.
-        batch_size
-            Minibatch size to use during training.
-        """
-        if max_epochs is None:
-            n_cells = self.adata.n_obs
-            max_epochs = np.min([round((20000 / n_cells) * 400), 400])
+    #         (loss, (elbo, new_model_state)), grads = jax.value_and_grad(
+    #             loss_fn, has_aux=True
+    #         )(state.params)
+    #         new_state = state.apply_gradients(
+    #             grads=grads, batch_stats=new_model_state["batch_stats"]
+    #         )
+    #         return new_state, loss, elbo, rngs
 
-        data_splitter = DataSplitter(
-            self.adata_manager,
-            train_size=train_size,
-            validation_size=validation_size,
-            batch_size=batch_size,
-            # for pinning memory only
-            use_gpu=False,
-            iter_ndarray=True,
-        )
-        data_splitter.setup()
-        train_loader = data_splitter.train_dataloader()
-        val_dataloader = data_splitter.val_dataloader()
+    #     @jax.jit
+    #     def validation_step(state, array_dict, rngs, **kwargs):
+    #         # note that self.module has is_training = False
+    #         module = self.module
+    #         rngs = {k: random.split(v)[1] for k, v in rngs.items()}
+    #         vars_in = {"params": state.params, "batch_stats": state.batch_stats}
+    #         outputs = module.apply(vars_in, array_dict, rngs=rngs, **kwargs)
+    #         loss_recorder = outputs[2]
+    #         loss = loss_recorder.loss
+    #         elbo = jnp.mean(loss_recorder.reconstruction_loss + loss_recorder.kl_local)
 
-        if val_dataloader is None:
-            raise UserWarning(
-                "No observations in the validation loop. No validation metrics will be recorded."
-            )
+    #         return loss, elbo
 
-        module_kwargs = self.module_kwargs.copy()
-        module_kwargs.update(dict(is_training=True))
-        train_module = self._get_module(module_kwargs)
+    #     history = dict(
+    #         elbo_train=[], loss_train=[], elbo_validation=[], loss_validation=[]
+    #     )
+    #     epoch = 0
+    #     with tqdm.trange(1, max_epochs + 1) as t:
+    #         try:
+    #             for i in t:
+    #                 epoch += 1
+    #                 epoch_loss = 0
+    #                 epoch_elbo = 0
+    #                 counter = 0
+    #                 for data in train_loader:
+    #                     kl_weight = min(1.0, epoch / 400.0)
+    #                     # gets new key for each epoch
+    #                     state, loss, elbo, self.rngs = train_step(
+    #                         state,
+    #                         data,
+    #                         self.rngs,
+    #                         loss_kwargs=dict(kl_weight=kl_weight),
+    #                     )
+    #                     epoch_loss += loss
+    #                     epoch_elbo += elbo
+    #                     counter += 1
+    #                 history["loss_train"] += [jax.device_get(epoch_loss) / counter]
+    #                 history["elbo_train"] += [jax.device_get(epoch_elbo) / counter]
+    #                 t.set_postfix_str(
+    #                     f"Epoch {i}, Elbo: {epoch_elbo / counter}, KL weight: {kl_weight}"
+    #                 )
 
-        # if key is generated on CPU, model params will be on CPU
-        # we have to pay the price of a JIT compilation though
-        if use_gpu is False:
-            key = jax.jit(lambda i: random.PRNGKey(i), backend="cpu")
-        else:
-            # dummy function
-            def key(i: int):
-                return random.PRNGKey(i)
+    #                 # validation loop
+    #                 if (
+    #                     check_val_every_n_epoch is not None
+    #                     and epoch % check_val_every_n_epoch == 0
+    #                 ):
+    #                     val_counter = 0
+    #                     val_epoch_loss = 0
+    #                     val_epoch_elbo = 0
+    #                     for data in val_dataloader:
+    #                         val_loss, val_elbo = validation_step(
+    #                             state,
+    #                             data,
+    #                             self.rngs,
+    #                             loss_kwargs=dict(kl_weight=kl_weight),
+    #                         )
+    #                         val_epoch_loss += val_loss
+    #                         val_epoch_elbo += val_elbo
+    #                         val_counter += 1
+    #                     history["loss_validation"] += [
+    #                         jax.device_get(val_epoch_loss) / val_counter
+    #                     ]
+    #                     history["elbo_validation"] += [
+    #                         jax.device_get(val_epoch_elbo) / val_counter
+    #                     ]
 
-        self.rngs = {
-            "params": key(0),
-            "dropout": key(1),
-            "z": key(2),
-        }
-        module_init = train_module.init(self.rngs, next(iter(train_loader)))
-        params = module_init["params"]
-        batch_stats = module_init["batch_stats"]
+    #         except KeyboardInterrupt:
+    #             logger.info(
+    #                 "Keyboard interrupt detected. Attempting graceful shutdown."
+    #             )
 
-        optimizer = optax.chain(
-            optax.additive_weight_decay(weight_decay=1e-6),
-            optax.adam(lr, eps=0.01),
-        )
+    #     self.train_state = state
+    #     self.params = state.params
+    #     self.batch_stats = state.batch_stats
+    #     self.history_ = {k: pd.DataFrame(v, columns=[k]) for k, v in history.items()}
+    #     self.is_trained_ = True
 
-        state = TrainState.create(
-            apply_fn=train_module.apply,
-            params=params,
-            tx=optimizer,
-            batch_stats=batch_stats,
-        )
-
-        @jax.jit
-        def train_step(state, array_dict, rngs, **kwargs):
-            rngs = {k: random.split(v)[1] for k, v in rngs.items()}
-
-            # batch stats can't be passed here
-            def loss_fn(params):
-                vars_in = {"params": params, "batch_stats": state.batch_stats}
-                outputs, new_model_state = state.apply_fn(
-                    vars_in, array_dict, rngs=rngs, mutable=["batch_stats"], **kwargs
-                )
-                loss_recorder = outputs[2]
-                loss = loss_recorder.loss
-                elbo = jnp.mean(
-                    loss_recorder.reconstruction_loss + loss_recorder.kl_local
-                )
-                return loss, (elbo, new_model_state)
-
-            (loss, (elbo, new_model_state)), grads = jax.value_and_grad(
-                loss_fn, has_aux=True
-            )(state.params)
-            new_state = state.apply_gradients(
-                grads=grads, batch_stats=new_model_state["batch_stats"]
-            )
-            return new_state, loss, elbo, rngs
-
-        @jax.jit
-        def validation_step(state, array_dict, rngs, **kwargs):
-            # note that self.module has is_training = False
-            module = self.module
-            rngs = {k: random.split(v)[1] for k, v in rngs.items()}
-            vars_in = {"params": state.params, "batch_stats": state.batch_stats}
-            outputs = module.apply(vars_in, array_dict, rngs=rngs, **kwargs)
-            loss_recorder = outputs[2]
-            loss = loss_recorder.loss
-            elbo = jnp.mean(loss_recorder.reconstruction_loss + loss_recorder.kl_local)
-
-            return loss, elbo
-
-        history = dict(
-            elbo_train=[], loss_train=[], elbo_validation=[], loss_validation=[]
-        )
-        epoch = 0
-        with tqdm.trange(1, max_epochs + 1) as t:
-            try:
-                for i in t:
-                    epoch += 1
-                    epoch_loss = 0
-                    epoch_elbo = 0
-                    counter = 0
-                    for data in train_loader:
-                        kl_weight = min(1.0, epoch / 400.0)
-                        # gets new key for each epoch
-                        state, loss, elbo, self.rngs = train_step(
-                            state,
-                            data,
-                            self.rngs,
-                            loss_kwargs=dict(kl_weight=kl_weight),
-                        )
-                        epoch_loss += loss
-                        epoch_elbo += elbo
-                        counter += 1
-                    history["loss_train"] += [jax.device_get(epoch_loss) / counter]
-                    history["elbo_train"] += [jax.device_get(epoch_elbo) / counter]
-                    t.set_postfix_str(
-                        f"Epoch {i}, Elbo: {epoch_elbo / counter}, KL weight: {kl_weight}"
-                    )
-
-                    # validation loop
-                    if (
-                        check_val_every_n_epoch is not None
-                        and epoch % check_val_every_n_epoch == 0
-                    ):
-                        val_counter = 0
-                        val_epoch_loss = 0
-                        val_epoch_elbo = 0
-                        for data in val_dataloader:
-                            val_loss, val_elbo = validation_step(
-                                state,
-                                data,
-                                self.rngs,
-                                loss_kwargs=dict(kl_weight=kl_weight),
-                            )
-                            val_epoch_loss += val_loss
-                            val_epoch_elbo += val_elbo
-                            val_counter += 1
-                        history["loss_validation"] += [
-                            jax.device_get(val_epoch_loss) / val_counter
-                        ]
-                        history["elbo_validation"] += [
-                            jax.device_get(val_epoch_elbo) / val_counter
-                        ]
-
-            except KeyboardInterrupt:
-                logger.info(
-                    "Keyboard interrupt detected. Attempting graceful shutdown."
-                )
-
-        self.train_state = state
-        self.params = state.params
-        self.batch_stats = state.batch_stats
-        self.history_ = {k: pd.DataFrame(v, columns=[k]) for k, v in history.items()}
-        self.is_trained_ = True
-
-        self.module_kwargs.update(dict(is_training=False))
-        self._module = None
-        self.bound_module = self.module.bind(
-            {"params": self.params, "batch_stats": self.batch_stats}, rngs=self.rngs
-        )
+    #     self.module_kwargs.update(dict(is_training=False))
+    #     self._module = None
+    #     self.bound_module = self.module.bind(
+    #         {"params": self.params, "batch_stats": self.batch_stats}, rngs=self.rngs
+    #     )
 
     def get_latent_representation(
         self,
