+++ scvi/module/base/__init__.py
@@ -4,6 +4,7 @@ from ._base_module import (
     LossRecorder,
     PyroBaseModuleClass,
 )
+from ._jax_module_wrapper import JaxModuleWrapper, BatchTrainState
 from ._decorators import auto_move_data
 
 __all__ = [
@@ -12,4 +13,6 @@ __all__ = [
     "PyroBaseModuleClass",
     "auto_move_data",
     "JaxBaseModuleClass",
+    "JaxModuleWrapper",
+    "BatchTrainState",
 ]
