+++ scvi/utils/__init__.py
@@ -1,5 +1,6 @@
 from ._attrdict import attrdict
 from ._docstrings import setup_anndata_dsp
 from ._track import track
+from ._jax import device_selecting_PRNGKey
 
-__all__ = ["track", "setup_anndata_dsp", "attrdict"]
+__all__ = ["track", "setup_anndata_dsp", "attrdict", "device_selecting_PRNGKey"]
